Rename tools to just the specific tool they handle
Rename database_handler -> search_handler
