+++ commec/screen.py
@@ -203,7 +203,7 @@ class Screen:
         Call hmmscan` and `check_biorisk.py` to add biorisk results to `screen_file`.
         """
         logging.debug("\t...running hmmscan")
-        self.database_tools.biorisk_db.screen()
+        self.database_tools.biorisk_db.search()
         logging.debug("\t...checking hmmscan results")
         check_biorisk(self.database_tools.biorisk_db.out_file,
                       self.database_tools.biorisk_db.db_directory)
@@ -214,7 +214,7 @@ class Screen:
         pathogen protein screening results to `screen_file`.
         """
         logging.debug("\t...running %s", self.params.config.search_tool)
-        self.database_tools.protein_db.screen()
+        self.database_tools.protein_db.search()
         if not self.database_tools.protein_db.check_output(): #os.path.exists(search_output):
             raise RuntimeError(f"Protein search failed and {self.database_tools.protein_db.out_file} was not created. Aborting.")
 
@@ -247,7 +247,7 @@ class Screen:
 
         # Only run new blastn search if there are no previous results
         if not self.database_tools.nucleotide_db.check_output():
-            self.database_tools.nucleotide_db.screen()
+            self.database_tools.nucleotide_db.search()
 
         if not self.database_tools.nucleotide_db.check_output():
             raise RuntimeError(f"Nucleotide search failed and {self.database_tools.nucleotide_db.out_file} was not created. Aborting.")
@@ -268,11 +268,11 @@ class Screen:
             return
 
         logging.debug("\t...running benign hmmscan")
-        self.database_tools.benign_hmm.screen()
+        self.database_tools.benign_hmm.search()
         logging.debug("\t...running benign blastn")
-        self.database_tools.benign_blastn.screen()
+        self.database_tools.benign_blastn.search()
         logging.debug("\t...running benign cmscan")
-        self.database_tools.benign_cmscan.screen()
+        self.database_tools.benign_cmscan.search()
 
         coords = pd.read_csv(sample_name + ".reg_path_coords.csv")
         benign_desc =  pd.read_csv(self.database_tools.benign_hmm.db_directory + "/benign_annotations.tsv", sep="\t")
